+++ examples/openai/csv_scraper_openai.py
@@ -1,25 +1,19 @@
-""" 
-Basic example of scraping pipeline using SmartScraper from text
+"""
+Basic example of scraping pipeline using CSVScraperGraph from CSV documents
 """
 
 import os
 from dotenv import load_dotenv
-from scrapegraphai.graphs import SmartScraperGraph
-from scrapegraphai.utils import prettify_exec_info
+import pandas as pd
+from scrapegraphai.graphs import CSVScraperGraph
+from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
 
 load_dotenv()
-
 # ************************************************
-# Read the text file
+# Read the csv file
 # ************************************************
 
-FILE_NAME = "inputs/plain_html_example.txt"
-curr_dir = os.path.dirname(os.path.realpath(__file__))
-file_path = os.path.join(curr_dir, FILE_NAME)
-
-# It could be also a http request using the request model
-with open(file_path, 'r', encoding="utf-8") as file:
-    text = file.read()
+text = pd.read_csv("inputs/username.csv")
 
 # ************************************************
 # Define the configuration for the graph
@@ -35,21 +29,25 @@ graph_config = {
 }
 
 # ************************************************
-# Create the SmartScraperGraph instance and run it
+# Create the CSVScraperGraph instance and run it
 # ************************************************
 
-smart_scraper_graph = SmartScraperGraph(
-    prompt="List me all the news with their description.",
-    source=text,
+csv_scraper_graph = CSVScraperGraph(
+    prompt="List me all the last names",
+    source=str(text),  # Pass the content of the file, not the file object
     config=graph_config
 )
 
-result = smart_scraper_graph.run()
+result = csv_scraper_graph.run()
 print(result)
 
 # ************************************************
 # Get graph execution info
 # ************************************************
 
-graph_exec_info = smart_scraper_graph.get_execution_info()
+graph_exec_info = csv_scraper_graph.get_execution_info()
 print(prettify_exec_info(graph_exec_info))
+
+# Save to json or csv
+convert_to_csv(result, "result")
+convert_to_json(result, "result")