+++ scrapegraphai/helpers/models_tokens.py
@@ -19,7 +19,11 @@ models_tokens = {
         "gpt-4-32k": 32768,
         "gpt-4-32k-0613": 32768,
     },
-
+    "azure": {
+        "gpt-3.5-turbo": 4096,
+        "gpt-4": 8192,
+        "gpt-4-32k": 32768
+    },
     "gemini": {
         "gemini-pro": 128000,
     },