+++ scrapegraphai/nodes/__init__.py
@@ -14,3 +14,4 @@ from .search_internet_node import SearchInternetNode
 from .generate_scraper_node import GenerateScraperNode
 from .search_link_node import SearchLinkNode
 from .robots_node import RobotsNode
+from .generate_answer_csv_node import GenerateAnswerCSVNode