+++ scrapegraphai/nodes/rag_node.py
@@ -90,6 +90,8 @@ class RAGNode(BaseNode):
         if isinstance(embedding_model, OpenAI):
             embeddings = OpenAIEmbeddings(
                 api_key=embedding_model.openai_api_key)
+        elif isinstance(embedding_model, AzureOpenAIEmbeddings):
+            embeddings = embedding_model
         elif isinstance(embedding_model, AzureOpenAI):
             embeddings = AzureOpenAIEmbeddings()
         elif isinstance(embedding_model, Ollama):
@@ -131,3 +133,4 @@ class RAGNode(BaseNode):
 
         state.update({self.output[0]: compressed_docs})
         return state
+